Testing Scanner module

+++ src/Scanner/Scanner.py
@@ -5,13 +5,19 @@ from datetime import time
 from utils.Utils import unzip, save, load
 
 
-def load_db(db_name, temp_folder='.'):
+def load_db(db_name, temp_folder='.', remove_after=True):
     s = Scanner()
     db_uncompresed = os.path.abspath(db_name.replace('.dbz', '.dbraw'))
+    print(db_uncompresed)
     if db_name.endswith('.dbz'):
-        unzip(db_name, temp_folder)
+        print(db_name)
+        print(os.path.dirname(db_name))
+        print(os.path.abspath(os.path.dirname(db_name)))
+
+        unzip(db_name, os.path.dirname(db_name))
         s.load(db_uncompresed)
-        os.remove(db_uncompresed)
+        if remove_after:
+            os.remove(db_uncompresed)
     else:
         s.load(db_name)
     return s
